# docs/openapi3_1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/controlla_disponibilita": {
      "post": {
        "summary": "Controlla Disponibilita",
        "operationId": "controlla_disponibilita_controlla_disponibilita_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/submit": {
      "post": {
        "summary": "Handle Form",
        "operationId": "handle_form_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Prenotazione"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "Autenticazione"
        ],
        "summary": "Login Stub",
        "operationId": "login_stub_login_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/user/{user_id}/": {
      "get": {
        "tags": [
          "Utenti"
        ],
        "summary": "Get User Stub",
        "operationId": "get_user_stub_user__user_id___get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reserve/{user_id}/": {
      "get": {
        "tags": [
          "Prenotazioni"
        ],
        "summary": "Get Reserve Stub",
        "operationId": "get_reserve_stub_reserve__user_id___get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reserve/{reserve_id}": {
      "put": {
        "tags": [
          "Prenotazioni"
        ],
        "summary": "Update Reserve Stub",
        "operationId": "update_reserve_stub_reserve__reserve_id__put",
        "parameters": [
          {
            "name": "reserve_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reserve Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Prenotazioni"
        ],
        "summary": "Delete Reserve Stub",
        "operationId": "delete_reserve_stub_reserve__reserve_id__delete",
        "parameters": [
          {
            "name": "reserve_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reserve Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Prenotazione": {
        "properties": {
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "cognome": {
            "type": "string",
            "title": "Cognome"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "telefono": {
            "type": "string",
            "title": "Telefono"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          },
          "checkin": {
            "type": "string",
            "title": "Checkin"
          },
          "checkout": {
            "type": "string",
            "title": "Checkout"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "cognome",
          "email",
          "telefono",
          "checkin",
          "checkout"
        ],
        "title": "Prenotazione"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}